backend/server/apps/users/views/token_obtain_pair_view.py
from apps.users.serializers import TokenObtainPairSerializer
from rest_framework_simplejwt.views import TokenObtainPairView



class CustomTokenObtainPairView(TokenObtainPairView):
    # Replace the serializer with your custom
    serializer_class = TokenObtainPairSerializer