backend/requirements.txt
django==3.1.2
djoser==2.0.5
djangorestframework==3.11.2 # https://github.com/sunscrapers/djoser/issues/541
markdown==3.3 
django-filter==2.4.0
django_cors_headers==3.5.0
psycopg2-binary>=2.8
black==20.8b1
Pillow==7.1.0
ipython==7.30.1
isort[requirements]==4.3.21
pylint==2.7.4
telethon==1.22.0
watchdog[watchmedo]==0.10.3
django-autoslug==1.9.8
python-telegram-bot==13.7
django-model-utils==4.1.1
sentry-sdk==1.4.3
celery==5.1.2
python-decouple==3.5
redis==3.5.3
gunicorn==20.1.0
djangorestframework-simplejwt==5.1.0
funcy==1.14